requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=0.20.0
tweepy>=4.15.0
apscheduler>=3.10.1
openai>=0.27.0
groq>=0.4.0

# Optional: for async support
aiohttp>=3.8.0

# Optional: For production secrets management
# AWS Secrets Manager: pip install boto3
# GCP Secret Manager: pip install google-cloud-secret-manager
# HashiCorp Vault: pip install hvac

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# pylint>=2.15.0
# mypy>=0.990
